feat: перенести пакет user в модуль server
