+++ shareIt-server/src/main/java/ru/practicum/shareit/server/user/UserRepository.java
@@ -1,7 +1,8 @@
-package ru.practicum.shareit.user;
+package ru.practicum.shareit.server.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.practicum.shareit.user.model.User;
+
+import ru.practicum.shareit.server.user.model.User;
 
 import java.util.Optional;
