+++ shareIt-server/src/main/java/ru/practicum/shareit/server/user/dto/UserDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.user.dto;
+package ru.practicum.shareit.server.user.dto;
 
 import lombok.Data;
